% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{AdaptixConnect}
\alias{AdaptixConnect}
\title{Creates an Adaptix Connection object to be supplied to API calls.}
\usage{
AdaptixConnect(url, api.key)
}
\arguments{
\item{url}{the base URL of the API.}

\item{api.key}{Your Adaptix API key.}
}
\value{
The Connection object (list).
}
\description{
Creates an Adaptix Connection object to be supplied to API calls.
}
\examples{
AdaptixConnect(url = "https://alpha.adaptix.io/api", api.key = "123456789abcdefgh")
}
