% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{AdaptixDeleteHTTPRequest}
\alias{AdaptixDeleteHTTPRequest}
\title{Basic http DELETE request to Adaptix.}
\usage{
AdaptixDeleteHTTPRequest(conn, url, verbose = F)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{url}{the complete URL on which to perform the HTTP operation.}

\item{verbose}{display HTTP operation details.}
}
\value{
the complete \code{httr} \code{response}.
}
\description{
Basic http DELETE request to Adaptix.
}
\examples{
AdaptixDeleteHTTPRequest(conn = conn, url = "https://alpha.adaptix.io/api/streams/123456abcdef", verbose = TRUE)
}
