% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{AdaptixDeleteStreamPoints}
\alias{AdaptixDeleteStreamPoints}
\title{Delete the points of an Adaptix stream}
\usage{
AdaptixDeleteStreamPoints(conn, streamURL = NULL, streamID = NULL,
  verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{streamURL}{the location of the stream, as a URL (as provided in the "href" field of API responses).}

\item{streamID}{the stream id (ignored if streamURL is provided)}

\item{verbose}{display HTTP operation details.}
}
\value{
the collection of points that were just deleted, as a list
}
\description{
Delete the points of an Adaptix stream
}
\examples{
AdaptixCreateStream(conn = conn, streamID = "123456abcdef", verbose = F)
}
