% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{AdaptixForecast}
\alias{AdaptixForecast}
\title{Perform forecast.}
\usage{
AdaptixForecast(conn, stream, from = NULL, to = NULL, span = NULL,
  history = NULL, aggregated = FALSE, individual = TRUE,
  number.of.scenarios = 0, rate = NULL, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream from which to retrieve the points.}

\item{from}{if not NULL, the starting date of the forecast. If NULL, defaults to the last point of the stream.}

\item{to}{if not NULL, the closing date of the forecast. required if span is NULL.}

\item{span}{a "span string" describing the forecast horizon. required if to is NULL.}

\item{aggregated}{if TRUE, the forecast will be provided as an aggregated forecast scenario, along with confidence margins.}

\item{individual}{if TRUE, the forecast will be provided as multiple plausible}

\item{number.of.scenarios}{if individual parameter is TRUE, restricts the number of scenarios that can be provided. If equal to 0, all scenarios deemed plausible by Adaptix will be provided.}

\item{rate}{desired frame rate of the forecast scnearios. Currently under development.}

\item{verbose}{display HTTP operation details.}

\item{window}{a "span string" describing a constraint on the data history to consider in the forecast.}
}
\value{
a data.frame with the requested collection of points.
}
\description{
Perform forecast.
}
\examples{
AdaptixForecast(conn = conn, stream = "123456abcdef", from = "2017-01-01", to = "2017-01-02", individual = T, number.of.scenarios = 3, verbose = F)
}
