% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{AdaptixGetAllStreamIds}
\alias{AdaptixGetAllStreamIds}
\title{Get a vector of all Adaptix stream ids}
\usage{
AdaptixGetAllStreamIds(conn, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{verbose}{display HTTP operation details.}
}
\value{
a vector of streams ids
}
\description{
Get a vector of all Adaptix stream ids
}
\examples{
AdaptixGetAllStreamIds(conn = conn, verbose = F)
}
