% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{AdaptixGetForecastScenarios}
\alias{AdaptixGetForecastScenarios}
\title{Converts scenarios in a forecast response to a comprehensive list structure of scenarios}
\usage{
AdaptixGetForecastScenarios(forecast.response)
}
\arguments{
\item{forecast.response}{A forecast response as returned by an AdaptixForecast call.}
}
\value{
transformed list of scenarios
}
\description{
Converts scenarios in a forecast response to a comprehensive list structure of scenarios
}
