% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{AdaptixGetHTTPRequest}
\alias{AdaptixGetHTTPRequest}
\title{Basic http GET request to Adaptix.}
\usage{
AdaptixGetHTTPRequest(conn, url, verbose = F)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{url}{the complete URL on which to perform the HTTP operation.}

\item{verbose}{display HTTP operation details.}
}
\value{
the complete \code{httr} \code{response}.
}
\description{
Basic http GET request to Adaptix.
}
\examples{
AdaptixGetHTTPRequest(conn = conn, url = "https://alpha.adaptix.io/api/streams", verbose = TRUE)
}
