% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{AdaptixGetLastPoints}
\alias{AdaptixGetLastPoints}
\title{Collects the last N points of an Adaptix stream.}
\usage{
AdaptixGetLastPoints(conn, stream, n = 1, verbose = F)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream from which to retrieve the points.}

\item{n}{the number of trailing points to collect.}

\item{verbose}{display HTTP operation details.}
}
\value{
a data.frame with the requested collection of points.
}
\description{
Collects the last N points of an Adaptix stream.
}
\examples{
AdaptixGetLastPoints(conn = conn, stream = "123456abcdef", n = 5, verbose = F)
}
