% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{AdaptixGetNotifications}
\alias{AdaptixGetNotifications}
\title{Collects the notifications of an Adaptix stream.}
\usage{
AdaptixGetNotifications(conn, stream, from = NULL, to = NULL,
  groupby = NULL, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream from which to retrieve the notifications.}

\item{from}{if not NULL, the starting date of the time window.}

\item{to}{if not NULL, the closing date of the time window.}

\item{verbose}{display HTTP operation details.}

\item{group.by}{if NULL, the raw notifications will be retrieved. If supplied as a "span" string describing a time window, the points will be averaged over consecutive windows of the specified length - see Adaptix API documentation.}
}
\value{
a data.frame with the requested collection of points.
}
\description{
Collects the notifications of an Adaptix stream.
}
\examples{
AdaptixGetNotifications(conn = conn, stream = "123456abcdef", group.by = "1h",  verbose = F)
}
