% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{AdaptixGetPoints}
\alias{AdaptixGetPoints}
\title{Collects the points of an Adaptix stream.}
\usage{
AdaptixGetPoints(conn, stream, from = NULL, to = NULL, groupby = NULL,
  aggregator = NULL, fill = NULL, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream from which to retrieve the points.}

\item{from}{if not NULL, the starting date of the time window.}

\item{to}{if not NULL, the closing date of the time window.}

\item{aggregator}{if NULL, the type of aggregation that will be processed over the aggregation windows (group.by parameter) - see Adaptix API documentation for details.}

\item{fill}{if NULL, missing data in the stream points will be filled according to the supplied supported operator}

\item{verbose}{display HTTP operation details.}

\item{group.by}{if NULL, the raw points will be retrieved. If supplied as a "span" string describing a time window, the points will be retrieved as an aggregation (MEAN by default, for other options see aggregator parameter) over consecutive windows of the specified length - see Adaptix API documentation.}
}
\value{
a data.frame with the requested collection of points.
}
\description{
Collects the points of an Adaptix stream.
}
\examples{
AdaptixGetPoints(conn = conn, stream = "123456abcdef", group.by = "1h", aggregator = "MEAN", verbose = F)
}
