% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{AdaptixGetPredictionScenarios}
\alias{AdaptixGetPredictionScenarios}
\title{Converts scenarios in a prediction response to a comprehensive list structure of scenarios}
\usage{
AdaptixGetPredictionScenarios(prediction.response)
}
\arguments{
\item{prediction.response}{A prediction response as returned by an AdaptixPrediction call.}
}
\value{
transformed list of scenarios
}
\description{
Converts scenarios in a prediction response to a comprehensive list structure of scenarios
}
