% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{AdaptixGetStream}
\alias{AdaptixGetStream}
\title{Get a specific Adaptix stream}
\usage{
AdaptixGetStream(conn, streamURL = NULL, streamID = NULL, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{streamURL}{the location of the stream, as a URL (as provided in the "href" field of API responses).}

\item{streamID}{the stream id (ignored if streamURL is provided)}

\item{verbose}{display HTTP operation details.}
}
\value{
the stream information, as a list
}
\description{
Get a specific Adaptix stream
}
\examples{
AdaptixGetStream(conn = conn, streamID = "123456abcdef", verbose = F)
AdaptixGetStream(conn = conn, streamURL = "https://alpha.adaptix.io/api/streams/123456abcdef", verbose = F)
}
