% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{AdaptixGetStreams}
\alias{AdaptixGetStreams}
\title{List all Adaptix streams.}
\usage{
AdaptixGetStreams(conn, as.data.frame = T, expand = F, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{as.data.frame}{if TRUE, returns result as a data.frame object, otherwise as a list.}

\item{expand}{wether or not to request the full (expanded) information of the requested resource(s) - see Adaptix API documentation.}

\item{verbose}{display HTTP operation details.}
}
\value{
a listings of streams.
}
\description{
List all Adaptix streams.
}
\examples{
AdaptixGetStreams(conn = conn, verbose = F)
}
