% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{AdaptixGetStreamsWithNames}
\alias{AdaptixGetStreamsWithNames}
\title{Get a dataframe of all Adaptix streams as a data frame with stream id, name, and href columns}
\usage{
AdaptixGetStreamsWithNames(conn, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{verbose}{display HTTP operation details.}
}
\value{
a listing of streams, in a dataframe format.
}
\description{
Get a dataframe of all Adaptix streams as a data frame with stream id, name, and href columns
}
\examples{
AdaptixGetStreamsWithNames(conn = conn, verbose = F)
}
