% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{AdaptixPredict}
\alias{AdaptixPredict}
\title{Perform prediction}
\usage{
AdaptixPredict(conn, stream, from = NULL, to = NULL, at = NULL, rate,
  seasonality, filter = NULL, rebuild = FALSE, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream on which to perform the prediction.}

\item{from}{if not NULL, the starting date of the prediction. If NULL, defaults to the last point of the stream.}

\item{to}{if not NULL, the closing date of the prediction. required if 'at' is NULL.}

\item{at}{if not NULL, the date of the prediction. required if 'from' and 'to' are NULL.}

\item{rate}{desired frame rate of the prediction scneario(s).}

\item{seasonality}{expected seasonality of the predicted data.}

\item{filter}{filter the data used in prediction to provided labels. Should be provided as a list.}

\item{verbose}{display HTTP operation details.}
}
\value{
a data.frame with the requested collection of points.
}
\description{
Perform prediction
}
\examples{
AdaptixPredictions(conn = conn, stream = "123456abcdef", from = "2017-01-01", to = "2017-01-02", seasonality = "1w", rate = "1h", verbose = F)
}
