% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{AdaptixPublishPoints}
\alias{AdaptixPublishPoints}
\title{Publishes points to an Adaptix stream.}
\usage{
AdaptixPublishPoints(conn, stream, points, group.by = 1, verbose = F)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream from which to retrieve the points.}

\item{points}{a data frame with "at" and "value" columns containing the points to publish to Adaptix.}

\item{group.by}{chunk the operation in several requests, each publiching "group.by" points of the supplied data frame.}

\item{verbose}{display HTTP operation details.}
}
\value{
the URL location of the created points
}
\description{
Publishes points to an Adaptix stream.
}
\examples{
AdaptixPublishPoints(conn = conn, stream = "123456abcdef", points = my_points.df, verbose = F)
}
