% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{AdaptixPutHTTPRequest}
\alias{AdaptixPutHTTPRequest}
\title{Basic http PUT request to Adaptix.}
\usage{
AdaptixPutHTTPRequest(conn, url, payload, verbose = F)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{url}{the complete URL on which to perform the HTTP operation.}

\item{payload}{payload of the request, supplied either as a JSON \code{character} string or a \code{list}.}

\item{verbose}{display HTTP operation details.}
}
\value{
the complete \code{httr} \code{response}.
}
\description{
Basic http PUT request to Adaptix.
}
\examples{
AdaptixPutHTTPRequest(conn = conn, url = "https://alpha.adaptix.io/api/streams", payload =  list(name = "my_stream"), verbose = TRUE)
}
