% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{AdaptixSearchPattern}
\alias{AdaptixSearchPattern}
\title{Perform a pattern search.}
\usage{
AdaptixSearchPattern(conn, stream, from, to, similarity.ratio = 1,
  historical.depth.from = NULL, historical.depth.to = NULL,
  in.streams = NULL, verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{stream}{the ID of the stream from which to retrieve the points.}

\item{from}{if not NULL, the starting date of the pattern.}

\item{to}{if not NULL, the closing date of the pattern.}

\item{similarity.ratio}{the similarity ratio of the pattern search [0.0 to 1.0].}

\item{historical.depth.from}{currently ignored}

\item{historical.depth.to}{currently ignored}

\item{in.streams}{if not NULL, the pattern extracted form the specified "stream" will also be searched in other streams, identified by their ids (provide as a vector of ids)}

\item{verbose}{display HTTP operation details.}
}
\value{
a data.frame with the requested collection of points.
}
\description{
Perform a pattern search.
}
\examples{
AdaptixSearchPattern(conn = conn, stream = "123456abcdef", from = "2017-01-01", to = "2017-01-02", similarity.ratio = 0.95, verbose = F)
}
