% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{AdaptixUpdateStream}
\alias{AdaptixUpdateStream}
\title{Update an Adaptix stream}
\usage{
AdaptixUpdateStream(conn, streamURL = NULL, streamID = NULL, name = NULL,
  device = NULL, description = NULL, units = NULL,
  anomaly.monitoring = F, store.points = T, data.rate = NULL,
  verbose = FALSE)
}
\arguments{
\item{conn}{a valid Adaptix connection object.}

\item{streamURL}{the location of the stream, as a URL (as provided in the "href" field of API responses).}

\item{streamID}{the stream id (ignored if streamURL is provided)}

\item{name}{the name of the stream.}

\item{device}{if not NULL, associates the stream with a device.}

\item{description}{the name of the stream.}

\item{units}{if not NULL, documents the unit ofor this stream's data.}

\item{anomaly.monitoring}{[REQUIRES ANOMALY DETECTION APP TO BE ENABLED] Enable anomaly detection on this stream.}

\item{store.points}{wether or not to store the points}

\item{data.rate}{if not NULL, indicates the normal data rate for the points of this stream.}

\item{verbose}{display HTTP operation details.}
}
\value{
the newly created stream information, as a list
}
\description{
Update an Adaptix stream
}
\examples{
AdaptixUpdateStream(conn = conn, name = "my_stream", verbose = F)
}
