% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ConvertDateToISO8601}
\alias{ConvertDateToISO8601}
\title{Converts a date to a formatted ISO8601 date}
\usage{
ConvertDateToISO8601(date, tz = "GMT")
}
\arguments{
\item{date}{a date, either as \code{character} string or \code{POSIXct} object.}

\item{tz}{the timezone of the date (relevant if only when date is \code{POSIXct}).}
}
\value{
an ISO8601 formatted date character string .
}
\description{
Converts a date to a formatted ISO8601 date
}
\examples{
ConvertDateToISO8601(date = as.POSIXct("2017-01-01"), tz = "GMT")
ConvertDateToISO8601(date = "2017-01-01")
}
