% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ConvertISO8601ToDate}
\alias{ConvertISO8601ToDate}
\title{Converts a formatted ISO8601 date to a POSIXct.}
\usage{
ConvertISO8601ToDate(iso8601, tz = "GMT")
}
\arguments{
\item{iso8601}{a ISO8601 date (please note: ISO8601 dates are GMT).}

\item{tz}{the timezone in which the output date should be issued.}
}
\value{
an ISO8601 formatted date character string .
}
\description{
Converts a formatted ISO8601 date to a POSIXct.
}
\examples{
ConvertISO8601ToDate(iso8601 = "2016-12-31T23:00:00.000Z", tz = "GMT")
}
