% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ConvertSpanStringToSeconds}
\alias{ConvertSpanStringToSeconds}
\title{A utility function to convert "span strings"}
\usage{
ConvertSpanStringToSeconds(span)
}
\arguments{
\item{span}{a time span string in the form of [amount|unit]. e.g. "1h", "35m", "3d"...}
}
\value{
the number of seconds corresponding to the supplied time span string
}
\description{
A utility function to convert "span strings"
}
\examples{
ConvertSpanStringToSeconds(span = "2h")
}
