% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{AdaptixForecastPointsToDataframe}
\alias{AdaptixForecastPointsToDataframe}
\title{Convert a list returned by AdaptixForecast call to a data frame}
\usage{
AdaptixForecastPointsToDataframe(forecast.list)
}
\arguments{
\item{forecast.list}{.}
}
\value{
.
}
\description{
Convert a list returned by AdaptixForecast call to a data frame
}
