% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{AdaptixForecastPointsToXts}
\alias{AdaptixForecastPointsToXts}
\title{Convert a list returned by AdaptixForecast call to an Xts object}
\usage{
AdaptixForecastPointsToXts(forecast.list)
}
\arguments{
\item{forecast.list}{.}
}
\value{
.
}
\description{
Convert a list returned by AdaptixForecast call to an Xts object
}
